+++ ide/web/lib/git_salt/git_salt.dart
@@ -5,6 +5,8 @@
 
 library spark.gitsalt;
 
+import 'package:chrome/chrome_app.dart' as chrome;
+
 import 'dart:async';
 import 'dart:js' as js;
 
@@ -36,6 +38,7 @@ class GitSalt {
   js.JsObject _jsGitSalt;
   int messageId = 1;
   String url;
+  chrome.DirectoryEntry root;
   Completer _completer = null;
 
   GitSalt() {
@@ -75,9 +78,8 @@ class GitSalt {
   bool get isActive => _completer != null;
 
   Future clone(entry, String url) {
-    if (isActive) {
-      return new Future.error("Another git operation in progress.");
-    }
+
+    root = entry;
 
     var arg = new js.JsObject.jsify({
       "entry": entry.toJs(),
@@ -97,19 +99,9 @@ class GitSalt {
     return _completer.future;
   }
 
-  void commitCb(var result) {
-    //TODO(grv): to be implemented.
-    print("commit successful");
-  }
+  Future commit(Map options) {
 
-  void commit(entry, String url) {
-
-    var arg = new js.JsObject.jsify({
-      "entry": entry.toJs(),
-      "filesystem": entry.filesystem.toJs(),
-      "fullPath": entry.fullPath,
-      "url": url
-    });
+    var arg = new js.JsObject.jsify(options);
 
     var message = new js.JsObject.jsify({
       "subject" : genMessageId(),
@@ -117,7 +109,15 @@ class GitSalt {
       "arg": arg
     });
 
-    _jsGitSalt.callMethod('postMessage', [message, commitCb]);
+    Completer completer = new Completer();
+
+    Function cb = (result) {
+      completer.complete();
+    };
+
+    _jsGitSalt.callMethod('postMessage', [message, cb]);
+
+    return completer.future;
   }
 
   Future<String> getCurrentBranch() {
@@ -166,6 +166,56 @@ class GitSalt {
 
     Function cb = (result) {
       completer.complete(result["branches"].toList());
+    };
+
+    _jsGitSalt.callMethod('postMessage', [message, cb]);
+
+    return completer.future;
+  }
+
+  Future add(List<chrome.Entry> entries) {
+
+    entries = entries.map((entry) {
+      if (entry.fullPath.length > root.fullPath.length) {
+        return entry.fullPath.substring(root.fullPath.length + 1);
+      } else {
+        return entry.fullPath;
+      }
+    }).toList();
+
+    var arg = new js.JsObject.jsify({
+      "entries" : entries
+    });
+
+    var message = new js.JsObject.jsify({
+      "subject" : genMessageId(),
+      "name" : "add",
+      "arg": arg
+    });
+
+    Completer completer = new Completer();
+
+    Function cb = (result) {
+      completer.complete();
+    };
+
+    _jsGitSalt.callMethod('postMessage', [message, cb]);
+
+    return completer.future;
+  }
+
+  Future<Map<String, String>> status() {
+
+    var message = new js.JsObject.jsify({
+      "subject" : genMessageId(),
+      "name" : "status",
+      "arg": {}
+    });
+
+    Completer completer = new Completer();
+
+    Function cb = (result) {
+      completer.complete(result["statuses"]);
     };
 
     _jsGitSalt.callMethod('postMessage', [message, cb]);