+++ ide/web/lib/scm.dart
@@ -456,14 +456,35 @@ class GitSaltScmProjectOperations extends ScmProjectOperations {
     return new Future.value();
   }
 
+
   Future commit(String userName, String userEmail, String commitMessage) {
-    // TODO(grv): Implement.
-    return new Future.value();
+    return gitSalt.then((git_salt) {
+      Map<String, String> options = {
+        "commitMessage": commitMessage,
+        "userName": userName,
+        "userEmail": userEmail
+      };
+      return git_salt.commit(options).then((_) {
+        return _refreshStatus(project: project);
+      }).catchError((e) => throw SparkException.fromExcetption(e));
+    });
   }
 
   Future push(String username, String password) {
     // TODO(grv): Implement.
     return new Future.value();
+  }
+
+  Future<List<String>> getDeletedFiles() {
+    return new Future.value([]);
+  }
+
+  Future addFiles(List<chrome.Entry> files) {
+    return gitSalt.then((git_salt) {
+      return git_salt.add(files).then((_) {
+        return _refreshStatus(project: project);
+      });
+    });
   }
 
   Stream<ScmProjectOperations> get onStatusChange => _statusController.stream;
@@ -476,8 +497,54 @@ class GitSaltScmProjectOperations extends ScmProjectOperations {
   }
 
   Future _refreshStatus({Project project, Iterable<Resource> resources}) {
-    //TODO(grv): implement
-    return new Future.value();
+    assert(project != null || resources != null);
+
+    // Get a list of all files in the project.
+    if (project != null) {
+      resources = project.traverse();
+    }
+
+    return gitSalt.then((git_salt) {
+      if (project != null) {
+        return git_salt.status().then((Map<String, String> statuses) {
+          resources.forEach((resource) {
+            _setStatus(resource, statuses[resource.entry.fullPath]);
+          });
+          return new Future.value();
+        });
+      } else {
+          return git_salt.status().then((Map<String, String> statuses) {
+            resources.forEach((resource) {
+              String rootPath = resource.project.entry.fullPath;
+              String path = resource.entry.fullPath.substring(rootPath.length + 1);
+              //TODO(grv): Update status for ancestors.
+              _setStatus(resource, statuses[path]);
+          });
+          return new Future.value();
+        });
+      }
+    }).catchError((e, st) {
+      _logger.severe("error calculating scm status", e, st);
+    }).whenComplete(() => _statusController.add(this));
+  }
+
+  void _setStatus(Resource resource, String status) {
+    String fileStatus = FileStatusType.COMMITTED;
+    //TODO(grv): Add a type class for the  returned status types.
+    // Handle case of untracked files.
+    if (status == null) {
+      fileStatus = FileStatusType.COMMITTED;
+    } else if (resource.isFile) {
+      if (status == 256) {
+          fileStatus = FileStatusType.MODIFIED;
+      } else  if (status == 512) {
+        fileStatus = FileStatusType.DELETED;
+      } else if (status == 128) {
+        fileStatus = FileStatusType.ADDED;
+      }
+    }
+    resource.setMetadata('scmStatus', new ScmFileStatus.fromIndexStatus(
+        fileStatus).status);
   }
 }